Add new test assets for proofmode JSON

+++ lib/validate/proofmode.py
@@ -35,6 +35,8 @@ class ProofMode(Validate):
         return "proofmode"
 
     def validate(self) -> bool:
+        found_json = False
+
         with ZipFile(self.zip_path, "r") as zipf:
             # Get dearmored key
             dearmored_key_path = os.path.join(self.tmp_dir, "dearmored_key")
@@ -86,12 +88,18 @@ class ProofMode(Validate):
                         shutil.rmtree(self.tmp_dir)
                         return False
 
+                    found_json = True
+
                     # It validated, add it to the data
                     self.files[file_name] = {
                         "signature": read_file(sig_path),
                         "authenticatedMessage": sha256sum(data_path),
                         "authenticatedMessageDescription": self.auth_msg_desc,
                     }
+
+        if not found_json:
+            # Proofmode ZIP has no JSON metadata file and must be considered invalid
+            return False
 
         shutil.rmtree(self.tmp_dir)
         return True
@@ -133,7 +141,7 @@ class ProofMode(Validate):
         )
         if proc.returncode == 0:
             return True
-        if proc.returncode == 1:
+        if proc.returncode == 1 or proc.returncode == 2:
             return False
         # Some other unexpected return code, means an error has occured
         proc.check_returncode()
